pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite==0.22.1",
    "alembic==1.17.2",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "certifi==2025.11.12",
    "click==8.3.1",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi-cli==0.0.16",
    "fastapi-cloud-cli==0.6.0",
    "fastapi[standard]==0.124.4",
    "fastar==0.8.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "phonenumbers==9.0.20",
    "psycopg2==2.9.11",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.10.6",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "rich==14.2.0",
    "rich-toolkit==0.17.0",
    "rignore==0.7.6",
    "sentry-sdk==2.47.0",
    "shellingham==1.5.4",
    "sqlalchemy==2.0.45",
    "sqlmodel==0.0.27",
    "starlette==0.50.0",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.2",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
